Декомпозировал редьюсер, написал селекторы и подключил декомпозированные редьюсеры в приложение
