+++ src/components/app/app.jsx
@@ -4,8 +4,9 @@ import {Switch, Route, BrowserRouter} from 'react-router-dom';
 import {connect} from 'react-redux';
 import Main from '../main/main.jsx';
 import OfferDetails from '../offer-details/offer-details.jsx';
-import {ActionCreator} from '../../reducer.js';
-import {getCurrentOffers} from '../../utils.js';
+import {ActionCreator as AppActionCreator} from '../../reducer/app/app.js';
+import {getOffers, getCities} from '../../reducer/data/selectors.js';
+import {getSelectedOffer, getCity} from '../../reducer/app/selectors.js';
 
 class App extends React.PureComponent {
   _renderApp() {
@@ -13,6 +14,7 @@ class App extends React.PureComponent {
       offers,
       selectedOffer,
       city,
+      cities,
       cityChangeHandler
     } = this.props;
 
@@ -20,6 +22,7 @@ class App extends React.PureComponent {
       return (
         <Main
           offers={offers}
+          cities={cities}
           city={city}
           cityChangeHandler={cityChangeHandler}
         />
@@ -41,20 +44,19 @@ class App extends React.PureComponent {
 
   render() {
     const {offers, city} = this.props;
-    const currentOffers = getCurrentOffers(offers, city);
     return (
       <BrowserRouter>
         <Switch>
           <Route exact path="/">
             {this._renderApp()}
           </Route>
-          {/* <Route exact path="/dev-details">
+          <Route exact path="/dev-details">
             <OfferDetails
-              offers={currentOffers}
-              id={currentOffers[0].id}
+              offers={offers}
+              id={offers[0]}
               city={city}
             />
-          </Route> */}
+          </Route>
         </Switch>
       </BrowserRouter>
     );
@@ -63,20 +65,22 @@ class App extends React.PureComponent {
 
 App.propTypes = {
   city: PropTypes.object,
+  cities: PropTypes.array,
   cityChangeHandler: PropTypes.func.isRequired,
   selectedOffer: PropTypes.oneOfType([PropTypes.object, PropTypes.number]),
   offers: PropTypes.array.isRequired,
 };
 
 const mapStateToProps = (state) => ({
-  offers: state.offers,
-  selectedOffer: state.selectedOffer,
-  city: state.city
+  offers: getOffers(state),
+  cities: getCities(state),
+  selectedOffer: getSelectedOffer(state),
+  city: getCity(state)
 });
 
 const mapDispatchToProps = (dispatch) => ({
   cityChangeHandler(city) {
-    dispatch(ActionCreator.changeCity(city));
+    dispatch(AppActionCreator.changeCity(city));
   }
 });
