+++ src/components/main/main.jsx
@@ -5,15 +5,8 @@ import {getCurrentOffers} from '../../utils.js';
 import Places from '../places/places.jsx';
 import PlacesEmpty from '../places-empty/places-empty.jsx';
 
-const getCities = (offers) => {
-  const cities = offers.map((offer) => offer.city.name);
-  const set = new Set(cities);
-  return Array.from(set);
-};
-
-const Main = ({offers, city, cityChangeHandler}) => {
+const Main = ({offers, city, cities, cityChangeHandler}) => {
   const currentOffers = getCurrentOffers(offers, city);
-  const cities = getCities(offers);
   const isEmpty = currentOffers.length === 0;
   return (
     <div className="page page--gray page--main">
@@ -65,6 +58,7 @@ const Main = ({offers, city, cityChangeHandler}) => {
 
 Main.propTypes = {
   city: PropTypes.object,
+  cities: PropTypes.array,
   cityChangeHandler: PropTypes.func.isRequired,
   offers: PropTypes.array,
   // offers: PropTypes.arrayOf(PropTypes.shape({
