+++ src/reducer/data/selectors.js
@@ -1,0 +1,9 @@
+import NameSpace from '../name-space.js';
+
+export const getOffers = (state) => {
+  return state[NameSpace.DATA].offers;
+};
+
+export const getCities = (state) => {
+  return state[NameSpace.DATA].cities;
+};
